Renamed save_and_load to game_state
